% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imitation_model.R
\name{imitation_model}
\alias{imitation_model}
\title{Imitation Model For Nucleotide Sequences}
\usage{
imitation_model(
  number_of_sequences,
  sequences_length,
  percent_of_new_positions,
  nucleotides,
  number_of_clusters
)
}
\arguments{
\item{number_of_sequences}{Integer, number of sequences in each cluster}

\item{sequences_length}{Integer, the length of each sequence}

\item{percent_of_new_positions}{Double, percentage of nucleotides that will
be regenerated in each sequence}

\item{nucleotides}{Character, alphabet of symbols from which sequences are}

\item{number_of_clusters}{Integer, number of clusters into which the data
will be divided}
}
\value{
The function returns a list of nucleotide sequences.
}
\description{
The function creates a list of nucleotide sequences based on the
first generated sequence with specified parameters. Imitation model
consists of two parts. \strong{generate_positions()} function returns nucleotide
positions that will be regenerated, while the \strong{generate_sequences()}
function performs generation of sequences. As a result, the obtained
sequences will have some similarity to each other, leading them to be
clustered together.
}
\examples{
imitation_model(6, 15, 0.4, c("A", "T", "C", "G"), 3)
     [,1]              [,2]              [,3]
[1,] "GAATTGGACCTCACA" "TTTCGGCGTTATGAA" "ACTGTGTCTTTTATT"
[2,] "GCATTCGACCTCCGA" "ATTCGCCGTTATCAT" "ATAGTGTCTTGTGTT"
[3,] "GCATTAGAGCTCAGA" "ATTCGGCGTTACTAC" "ATCGTGTCTTGTATT"
[4,] "GAATTCGAGCTCCAA" "TTTCGACGTTAGTAC" "ACAGTGTCTTGCCTT"
[5,] "GAATTCGAACTCCAA" "GTTCGCCGTTAACAT" "ATGGTGTCTTCGATT"
[6,] "GCATTAGAACTCCGA" "GTTCGTCGTTAACAC" "AGAGTGTCTTTACTT"
}
