% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaguchiClustering.R
\name{TaguchiClustering}
\alias{TaguchiClustering}
\title{Taguchi Clustering Method}
\usage{
TaguchiClustering(
  sequences,
  method = "lv",
  number_of_clusters = 4,
  PC_num = 10,
  clusters_plot = FALSE,
  new_clusters_plot = FALSE
)
}
\arguments{
\item{sequences}{List, non-vectorized nucleotide sequences}

\item{method}{Method for distance calculation. The default is "lv", but you can
also use another method: \emph{'osa'} – optimal alignment, \emph{'dl'} –
Damerau-Levenshtein, \emph{'qgram'} – q gram, \emph{'cosine'} – cosine, \emph{'jw'} – Jaro,
\emph{'lcs'} – greatest common substrings}

\item{number_of_clusters}{Integer, number of clusters which you want to divide
your sequences. The default is 4}

\item{PC_num}{Integer, number of principal coordinates u want to take from the
function}

\item{clusters_plot}{Bool, displays a plot on the first two principal coordinates}

\item{new_clusters_plot}{Bool, displays a plot on the first two principal
coordinates in new feature space}
}
\value{
List of \emph{P_num} principal coordinates, \emph{P_num} coordinates in the new
feature space
}
\description{
The function takes a list of sequences and a distance metric,
based on which the given sequences will be vectorized. Then the function
computes and returns the \emph{PC_num} number of principal coordinates and also
performs a transition to a new feature space.
}
\examples{
TaguchiClustering(imitation_model(), number_of_clusters = 8)
TaguchiClustering(imitation_model(), method = 'qgram', new_clusters_plot = TRUE)
}
